# Conflicts:
#   tingting/src/main/java/com/alsif/tingting/book/controller/BookController.java
#   tingting/src/main/java/com/alsif/tingting/book/dto/TicketBaseDto.java
#   tingting/src/main/java/com/alsif/tingting/book/repository/TicketRepository.java
#   tingting/src/main/java/com/alsif/tingting/book/repository/TicketSeatRepository.java
#   tingting/src/main/java/com/alsif/tingting/book/service/BookService.java
#   tingting/src/main/java/com/alsif/tingting/concert/controller/ConcertController.java
#   tingting/src/main/java/com/alsif/tingting/concert/dto/concerthall/SeatBaseDto.java
#   tingting/src/main/java/com/alsif/tingting/concert/repository/ConcertRepository.java
#   tingting/src/main/java/com/alsif/tingting/global/config/CorsConfig.java
#   tingting/src/main/java/com/alsif/tingting/user/service/UserService.java

+++ tingting/src/main/java/com/alsif/tingting/book/controller/BookController.java
@@ -136,7 +136,7 @@ public class BookController {
 	})
 	@PostMapping("/{concertDetailSeq}/seat")
 	ResponseEntity<SuccessResponseDto> book(
-		@PathVariable("concertDetailSeq") Integer concertDetailSeq, @RequestParam Integer userSeq,
+		@PathVariable("concertDetailSeq") Long concertDetailSeq, @RequestParam Long userSeq,
 		@RequestBody ConcertSeatBookRequestDto requestDto) {
 		log.info("===== 선택 좌석 예매 요청 시작, url={}, concertDetailSeq: {}, {} =====",
 			"/concerts", concertDetailSeq, requestDto.toString());
